+++ src/Trail.hpp
@@ -3,3 +3,4 @@
 void updateTrail(LevelEditorLayer* editor);
 void startTrail();
 void stopTrail();
+void toggleTrail(bool on);